Simple shell-based tests
After a new app snaphot is built by Packer, the CI will now run test.sh
to verify the image looks okay.

Since the internal CI looks for snapshots with the app= label, that
label is not applied until after the tests pass. Put differently, the
internal CI can now be run on a schedule without risk of publishing an
app release that has not yet been "blessed" from a successful test run.