[WireGuard] Update wireguard-ui and caddy
Because all required wireguard-ui patches were merged and released now,
we can switch back to upstream instead of my fork.
Also, a recent caddy release brought a breaking change in the behavior
of `caddy hash-password`, so the install script and README was updated
to reflect that. But there are upcoming changes in wireguard-ui, that
will add an UI way to change the password. When this is merged, we
can remove the command from the README.